+++ lib/src/vector/view.dart
@@ -1,3 +1,4 @@
+export 'view/binary_operation_vector.dart' show BinaryOperationVectorExtension;
 export 'view/cast_vector.dart' show CastVectorExtension;
 export 'view/column_matrix.dart' show ColumnMatrixExtension;
 export 'view/diagonal_matrix.dart' show DiagonalMatrixExtension;
@@ -9,5 +10,6 @@ export 'view/range_vector.dart' show RangeVectorExtension;
 export 'view/reversed_vector.dart' show ReversedVectorExtension;
 export 'view/row_matrix.dart' show RowMatrixExtension;
 export 'view/transformed_vector.dart' show TransformedVectorExtension;
+export 'view/unary_operation_vector.dart' show UnaryOperationVectorExtension;
 export 'view/unmodifiable_vector.dart' show UnmodifiableVectorExtension;
 export 'view/vector_list.dart' show VectorListExtension;